changed return type to troubleshoot

+++ Temp_Sensor/temp_sens.cpp
@@ -2,6 +2,7 @@
  * Read Temperature from TMP0006 sensor
  */
 #include "temp_sens.h"
+#include "TMP006.h"
 
 I2C i2c(p9, p10);
 
@@ -10,8 +11,8 @@ const int addr = 0x80;
 //TMP0006 sends a 16bit (2 byte) signed value
 //LSB represents 1/32 Celcius
 char buff[2]; 
-/*
-float temp_Celcius() {
+
+int temp_Celcius() {
   char reg[1];
   char data[2];
   
@@ -35,9 +36,8 @@ float temp_Celcius() {
 	
 	//shifts top byte left by 8, then inclusive or with bottom
     //then it is forced into a float and divided by 32 (LSB is 1/32 Celcius)
-	return float( ( buff[0] << 8 ) | buff[1] ) / 32.0;
+	return int(( ( buff[0] << 8 ) | buff[1] ) / 32.0);
 }
-*/
 float temp_Fareinheit() {
     return 6.9;
 }
